+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 ## [0.2.0.dev0] - Unreleased
 - Updated documentation.
 - Fixed legend issue with rolling plots when showing only one horizon.
+- Added filtering of outturns in the filtering methods.