+++ CHANGELOG.md
@@ -3,8 +3,10 @@
 ## [0.1.0] - 2026-01-22
 - Initial release.
 
-## [0.2.0.dev0] - Unreleased
+## [0.1.1] - 2026-02-26
 - Updated documentation.
 - Fixed legend issue with rolling plots when showing only one horizon.
 - Added filtering of outturns in the filtering methods.
 - Removed method-chaining with filter() and merge() methods.
+
+## [0.1.2.dev0] - Unreleased